requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==5.1.1
executing==1.2.0
idna==3.4
ipython==8.11.0
jedi==0.18.2
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
tweepy==4.13.0
urllib3==1.26.15
wcwidth==0.2.6
